% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{bf_selection}
\alias{bf_selection}
\title{Compute the pairwise Bayes Factors between models}
\usage{
bf_selection(models)
}
\arguments{
\item{models}{A list of biPOD object of class \code{bipod} that have been fitted with the same prior.}
}
\value{
A tibble with pairwise Bayes Factors. The column names are the sample names of the models.
}
\description{
Compute the pairwise Bayes Factors between models
}
