% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biPOD-package.R
\docType{package}
\name{biPOD-package}
\alias{biPOD}
\alias{biPOD-package}
\title{biPOD: biPOD}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Bayesian inference of kinetic parameters (growth rates, birth time, change points) from time-series population count data, using Stan via 'cmdstanr' for MCMC and variational inference with model selection via LOO-IC or BIC.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://caravagnalab.github.io/biPOD/}
}

}
\author{
\strong{Maintainer}: Giovanni Santacatterina \email{santacatterinagiovanni@gmail.com}

}
\keyword{internal}
