% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc_plots.R
\name{biPOD_ppc_intervals}
\alias{biPOD_ppc_intervals}
\title{Posterior Predictive Intervals Plot}
\usage{
biPOD_ppc_intervals(x, x_fit, n = 500)
}
\arguments{
\item{x}{A \code{bipod} object that contains a 'counts' field with the observed data.}

\item{x_fit}{A fitted model object that contains posterior draws of the replicated data (yrep).}

\item{n}{Number of posterior predictive draws to use in the plot. (default is 500)}
}
\value{
A plot showing posterior predictive intervals using \code{bayesplot::ppc_intervals}.
}
\description{
Creates a plot displaying the posterior predictive intervals for the observed data.
This helps visualize how well the model captures the variation in the data over time.
}
