% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{birth_posterior_plot}
\alias{birth_posterior_plot}
\title{Plot the posterior distribution for the birth rates.}
\usage{
birth_posterior_plot(fit_model)
}
\arguments{
\item{fit_model}{Is the result obtained by fitting a stan model over the data.}
}
\value{
A plot. Represents the posterior distribution for the birth rates.
}
\description{
Plot the posterior distribution for the birth rates.
}
