% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnostics.R
\name{diagnose_chains}
\alias{diagnose_chains}
\title{Function to diagnose chains for a bipod object}
\usage{
diagnose_chains(x, pars = c(), rhat_threshold = 1.01)
}
\arguments{
\item{x}{a bipod object}

\item{pars}{a character vector of parameter names to be examined}

\item{rhat_threshold}{Threshold value accepted for Rhat value}
}
\value{
a logical vector indicating whether the chains for each fit passed or failed the diagnostic test
}
\description{
Function to diagnose chains for a bipod object
}
