% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{evolution_plot}
\alias{evolution_plot}
\title{Plot the evolution of the population over time}
\usage{
evolution_plot(d, estimate_rates = FALSE)
}
\arguments{
\item{d}{Data frame. Must have at least three columns with the following names:
"time", "count" or "pop.size", and "sample".}

\item{estimate_rates}{Boolean. If TRUE, curves computed by a rough estimation of the
growth rates for each time interval will be plotted.}
}
\value{
A plot. Represents the evolution of the population over time.
}
\description{
Plot the evolution of the population over time
}
