% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_exp}
\alias{fit_exp}
\title{Fit a Bayesian model for the evolution of a population.}
\usage{
fit_exp(
  x,
  factor_size = 1,
  gaussian_approx = F,
  rates_mode = "fixed_all",
  chains = 4,
  iter = 4000,
  warmup = 2000,
  cores = 4
)
}
\arguments{
\item{x}{A biPOD object of class \code{bipod}.}

\item{factor_size}{Integer. Counts will be divided by this factor size.}

\item{gaussian_approx}{Boolean. If True, a Gaussian approximation will be applied}

\item{rates_mode}{Boolean. If True, a Gaussian approximation will be applied
to the distributions used for the inference.}

\item{chains}{Integer. Number of MCMC chains}

\item{iter}{Integer. Number of MCMC steps}

\item{warmup}{Integer. Number of MCMC warmup steps, must be smaller than 'chains'.}

\item{cores}{Integer. Number of cores to use.}
}
\value{
A plot. Represents the evolution of the population over time.
}
\description{
Fit a Bayesian model for the evolution of a population.
}
