% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_log}
\alias{fit_log}
\title{Fit a  Bayesian model for the evolution of a population
with the assumption of a logistic growth.}
\usage{
fit_log(
  x,
  model_name,
  factor_size = 1,
  fix_rates = 0,
  prior_par = 2,
  chains = 4,
  iter = 4000,
  warmup = 2000,
  cores = 4
)
}
\arguments{
\item{x}{A biPOD object of class \code{bipod}.}

\item{model_name}{String. Indicates the name of the model desired.
\itemize{
\item \code{gauss} Use a gaussian approximation. Works well when counts data are far from 0.
}}

\item{factor_size}{Integer. Counts will be divided by this factor size.}

\item{fix_rates}{An integer value. Indicates which rates must be fixed.
\itemize{
\item \code{1} Growth rate is not fixed at every time step.
\item \code{0} Growth rate is fixed throughout all time steps.
}}

\item{prior_par}{Parameter for inverse gamma prior. InvGamma(p,p).}

\item{chains}{Integer. Number of MCMC chains}

\item{iter}{Integer. Number of MCMC steps}

\item{warmup}{Integer. Number of MCMC warmup steps, must be smaller than 'chains'.}

\item{cores}{Integer. Number of cores to use.}
}
\value{
A plot. Represents the evolution of the population over time.
}
\description{
Fit a  Bayesian model for the evolution of a population
with the assumption of a logistic growth.
}
