% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{fit_plot}
\alias{fit_plot}
\title{Plot the fit over the input data.}
\usage{
fit_plot(x)
}
\arguments{
\item{x}{A biPOD object of class \code{bipod}. Must contains 'fit'}
}
\value{
A plot of the fit over the input data.
}
\description{
Plot the fit over the input data.
}
