% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_task3.R
\name{fit_two_pop_model}
\alias{fit_two_pop_model}
\title{Fit a Two-Population Growth Model to a \code{bipod} Object}
\usage{
fit_two_pop_model(
  x,
  variational = FALSE,
  factor_size = 1,
  chains = 4,
  iter = 5000,
  cores = 4
)
}
\arguments{
\item{x}{A \code{bipod} object containing population count data over time.}

\item{variational}{Logical, indicating whether to use Variational Inference instead of MCMC sampling.
If \code{TRUE}, Variational Inference is applied; otherwise, MCMC sampling is used. Default is \code{FALSE}.}

\item{factor_size}{Numeric, specifying the scaling factor for population counts in the \code{bipod} object.
This value must be positive and appropriately chosen for the data. Default is \code{1}.}

\item{chains}{Integer, defining the number of MCMC chains. This parameter is ignored if \code{variational = TRUE}. Default is \code{4}.}

\item{iter}{Integer, setting the number of iterations for MCMC sampling. This parameter is ignored if \code{variational = TRUE}. Default is \code{5000}.}

\item{cores}{Integer, specifying the number of CPU cores to use for parallel computation. Default is \code{4}.}
}
\value{
The input \code{bipod} object with additional attributes:
\itemize{
\item \code{two_pop_fit} – The fitted two-population growth model.
\item \code{two_pop_fit_info} – Metadata related to the fitting process, including:
\itemize{
\item Sampling method (MCMC or Variational Inference)
\item Scaling factor used (\code{factor_size})
\item Diagnostics of the model fitting process
}
\item \code{two_pop_plots} – A list of visualization plots, including:
\itemize{
\item Evolutionary trend plots
\item Population growth rates
\item Key time estimates (e.g., emergence of resistant populations)
}
}
}
\description{
This function fits a growth model that accounts for two interacting populations within a \code{bipod} object.
The model can be fitted using either Variational Inference or Markov Chain Monte Carlo (MCMC) sampling.
}
\examples{
# Load a bipod object
data("xenografts", package = "biPOD")
mouse_id <- 543
d <- xenografts
d = dplyr::rename(d, count = tumour_volume)
d = dplyr::mutate(d, time = time / 7)
d = dplyr::filter(d, time >= 0)
d = dplyr::filter(d, mouse == mouse_id)
x <- biPOD::init(d, "543 U-shape")
# Fit a two-population growth model using MCMC
x <- biPOD::fit_two_pop_model(x, variational = FALSE, factor_size = 1)
biPOD::plot_two_pop_fit(x, split_process = FALSE, f_posteriors = FALSE, t_posteriors = FALSE, r_posteriors = FALSE)

}
