% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_posterior.R
\name{get_growth_rate_posteriors}
\alias{get_growth_rate_posteriors}
\title{Plot growth rates posteriors}
\usage{
get_growth_rate_posteriors(x, add_prior = F, labels = NULL)
}
\arguments{
\item{x}{a bipod object with a 'fit' field}

\item{add_prior}{Boolean, indicate whether to plot also the prior distribution}

\item{labels}{Vector of labels for the growth rate of each group. If NULL, the standard is 'rho' plus the group index.}
}
\value{
A ggplot object containing the posterior density plots of the growth rates and the prior density plot
}
\description{
Plot growth rates posteriors
}
