% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loo_selection.R
\name{loo_selection}
\alias{loo_selection}
\title{Compute the pairwise LOO between models}
\usage{
loo_selection(models)
}
\arguments{
\item{models}{A list of biPOD object of class \code{bipod} that have been fitted with the same prior.}
}
\value{
A tibble with pairwise LOOs. The column names are the sample names of the models.
}
\description{
Compute the pairwise LOO between models
}
