% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bayes_factor.R
\name{plot_bayes_factor}
\alias{plot_bayes_factor}
\title{Plot the fit over the input data.}
\usage{
plot_bayes_factor(x)
}
\arguments{
\item{x}{A biPOD object of class \code{bipod}. Must contains 'fit' and must have been fitted with model selection.}
}
\value{
A plot of the Bayes Factor with its significance.
}
\description{
Plot the fit over the input data.
}
