% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_posterior.R
\name{plot_birth_and_death_rates_posteriors}
\alias{plot_birth_and_death_rates_posteriors}
\title{Plot birth and death rates posteriors}
\usage{
plot_birth_and_death_rates_posteriors(
  x,
  point_est = c("mean", "median", "none")
)
}
\arguments{
\item{x}{a bipod object with a 'fit' field}

\item{point_est}{Character string indicating which point estimate to use. Available options are "mean", "median", and "none". Default is "mean".}
}
\value{
A ggplot object containing the posterior density plots of the birth and death rates and the prior density plot
}
\description{
Plot birth and death rates posteriors
}
