% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_breakpoints.R
\name{plot_breakpoint_posterior}
\alias{plot_breakpoint_posterior}
\title{Plot posterior distributions of inferred breakpoints}
\usage{
plot_breakpoint_posterior(x, colors = NULL, zoom = FALSE)
}
\arguments{
\item{x}{A biPOD object which must have passed through the \code{fit_breakpoints} function.}
}
\value{
A \code{ggplot} object showing the posterior densities for each breakpoint parameter.
}
\description{
Plots posterior densities of the inferred breakpoint times (\code{t_array[j]})
from a fitted breakpoint model.
}
