% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_breakpoints_posterior.R
\name{plot_breakpoints_posterior}
\alias{plot_breakpoints_posterior}
\title{Plot the posterior distributions of the inferred breakpoints.}
\usage{
plot_breakpoints_posterior(x)
}
\arguments{
\item{x}{a bipod object with a 'fit' field}
}
\value{
A ggplot object containing the posterior distributions of the inferred breakpoints.
}
\description{
Plot the posterior distributions of the inferred breakpoints.
}
