% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_evolution.R
\name{plot_evolution}
\alias{plot_evolution}
\title{Plot the evolution of the population over time}
\usage{
plot_evolution(x, add_title = F)
}
\arguments{
\item{x}{A biPOD object of class \code{bipod}.}

\item{add_title}{Boolean, indicating whether the plot should have a title}
}
\value{
A plot. Represents the evolution of the population over time.
}
\description{
Plot the evolution of the population over time
}
