% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fits.R
\name{plot_fit}
\alias{plot_fit}
\title{Plot the fit over the input data.}
\usage{
plot_fit(x)
}
\arguments{
\item{x}{A biPOD object of class \code{bipod}. Must contains 'fit'}
}
\value{
A plot of the fit over the input data.
}
\description{
Plot the fit over the input data.
}
