% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_posterior.R
\name{plot_growth_rate_posteriors}
\alias{plot_growth_rate_posteriors}
\title{Plot growth rates posteriors}
\usage{
plot_growth_rate_posteriors(x)
}
\arguments{
\item{x}{a bipod object with a 'fit' field}
}
\value{
A ggplot object containing the posterior density plots of the growth rates and the prior density plot
}
\description{
Plot growth rates posteriors
}
