% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_evolution.R
\name{plot_input}
\alias{plot_input}
\title{Plot the input data of the population over time}
\usage{
plot_input(x, add_title = F, log_scale = F)
}
\arguments{
\item{x}{A biPOD object of class \code{bipod}.}

\item{add_title}{Boolean, indicating whether the plot should have a title}

\item{log_scale}{Boolean, indicating whether the plot should have a title}
}
\value{
A plot. Represents the evolution of the population over time.
}
\description{
Plot the input data of the population over time
}
