% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_simple_fit.R
\name{plot_simple_fit}
\alias{plot_simple_fit}
\title{Plot a simple fit over the input data.}
\usage{
plot_simple_fit(
  x,
  CI = 0.95,
  full_process = F,
  legend_labels = NULL,
  legend_title = "group",
  t0_posterior_color = "darkorange",
  shadows_colors = NULL
)
}
\arguments{
\item{x}{A bipod object of class \code{bipod}. Must contains 'fit'}

\item{CI}{confidence interval for the growth rate to plot}

\item{full_process}{Boolean, indicating whether to plot the process starting from t0 or not.}

\item{legend_labels}{Vector containing a name for each unique group in x$counts$group}

\item{legend_title}{Title for the legend. Default is "group"}

\item{t0_posterior_color}{.}

\item{shadows_colors}{colors to use for the different time windows}
}
\value{
A plot of the fit over the input data.
}
\description{
Plot a simple fit over the input data.
}
