% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_growth_threshold.R
\name{plot_threshold_crossing_time_distribution}
\alias{plot_threshold_crossing_time_distribution}
\title{Plot the the posterior distribution of the 'relapse' times.}
\usage{
plot_threshold_crossing_time_distribution(x, n_thresh)
}
\arguments{
\item{x}{A biPOD object of class \code{bipod}. Must contains 'fit'}

\item{n_thresh}{A positive value indicating the threshold for which
the posterior distribution must be computed}
}
\value{
A posterior distribution plot.
}
\description{
Plot the the posterior distribution of the 'relapse' times.
}
