% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnostics.R
\name{plot_traces}
\alias{plot_traces}
\title{Plot Traces of Specified Parameters from a Fitted Stan Model}
\usage{
plot_traces(x, fit, pars = c(), diagnose = FALSE)
}
\arguments{
\item{x}{A \code{bipod} object that contains at least one fitted model.}

\item{fit}{The fit within the \code{bipod} object for which the traces should be plotted.}

\item{pars}{A character vector specifying the names of the parameters to plot. If empty, plots all available parameters.}

\item{diagnose}{A logical value indicating whether to color the plots based on convergence diagnostics.
If \code{TRUE}, the plots will include color coding based on the R-hat convergence statistic. (default is FALSE)}
}
\value{
A \code{ggplot2} object displaying the trace plots for the specified parameters.
}
\description{
Generates trace plots for specified parameters from a fitted Stan model contained within a \code{bipod} object. The plots visualize the sampled values across iterations for each chain.
}
