% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots_growth_rate_posteriors.R
\name{plots_of_growth_rate_posteriors}
\alias{plots_of_growth_rate_posteriors}
\title{Produces a list of plots, representing the posteriors, one for each growth rate.}
\usage{
plots_of_growth_rate_posteriors(x)
}
\arguments{
\item{x}{a bipod object with a 'fit' field}
}
\value{
A ggplot object containing the posterior density plots of the growth rates and the prior density plot
}
\description{
Produces a list of plots, representing the posteriors, one for each growth rate.
}
