% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{posterior_plot}
\alias{posterior_plot}
\title{Plot the posterior distribution for a specific group of parameters.}
\usage{
posterior_plot(x, param_name)
}
\arguments{
\item{x}{A biPOD object of class \code{bipod}. Must contains 'fit'}

\item{param_name}{A string indicating the desired parameters.}
}
\value{
A plot. Represents the posterior distribution for the desired parameters.
}
\description{
Plot the posterior distribution for a specific group of parameters.
}
