% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ppc_plot}
\alias{ppc_plot}
\title{Plot the posterior predictive checks for counts data.}
\usage{
ppc_plot(x, ptype, prob = 0.5, prob_outer = 0.9)
}
\arguments{
\item{x}{A biPOD object of class \code{bipod}. Must contains 'fit'}

\item{ptype}{A string representing the type of visualization.

\code{intervals} vertical bars with points indicating generated values medians and darker points indicating observed values
\code{ribbon} ribbon of connected intervals with a line through the median of generated values and a darker line connecting observed values}

\item{prob, prob_outer}{Values between 0 and 1 indicating the desired probability mass to include in the inner and outer intervals. The defaults are prob=0.5 and prob_outer=0.9.}
}
\value{
A plot. Represents the posterior predictive checks.
}
\description{
Plot the posterior predictive checks for counts data.
}
