% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bipod.R
\name{print.bipod}
\alias{print.bipod}
\title{Print for class \code{'bipod'}.}
\usage{
\method{print}{bipod}(x, ...)
}
\arguments{
\item{x}{An obj of class \code{'bipod'}.}

\item{...}{Default S3 method parameter.}
}
\value{
Nothing.
}
\description{
Print for class \code{'bipod'}.
}
