% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_logistic.R
\name{sim_single_stochastic_logistic}
\alias{sim_single_stochastic_logistic}
\title{Simulate population growth with logistic growth}
\usage{
sim_single_stochastic_logistic(n0, lambda, mu, K, delta_t)
}
\arguments{
\item{n0}{Initial population size (must be 1).}

\item{lambda}{Growth rate.}

\item{mu}{Death rate.}

\item{K}{Carrying capacity.}

\item{delta_t}{Time step.}
}
\value{
The final population size.
}
\description{
This function simulates population growth for a single individual with a logistic
growth model. The population size at each time step is calculated based on the
growth rate (lambda) and death rate (mu) at that step, and the carrying capacity
(K). The time step (delta_t) is determined by randomly sampling from an exponential
distribution. The function returns the final population size.
}
