% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_stochastic}
\alias{sim_stochastic}
\title{Simulate a stochastic birth-death process for multiple time steps}
\usage{
sim_stochastic(n0, lambda, mu, steps, delta_t)
}
\arguments{
\item{n0}{Integer value greater than zero. Starting population.}

\item{lambda}{Vector of real values or single real value. Must be greater than zero. Birth rates.}

\item{mu}{Vector of real values or single real value. Must be greater than zero. Death rates.}

\item{steps}{Integer value strictly greater than zero. Number of time steps.}

\item{delta_t}{Vector of real values or single real value. Must be greater than zero. Time intervals.}
}
\value{
A data frame with the columns: "time", "count", "death.rates", and "birth.rates".
}
\description{
Simulate a stochastic birth-death process for multiple time steps
}
