\name{BIPOD-package}
\alias{BIPOD-package}
\docType{package}
\title{
Bayesian parameter estimation in two-dimensional diffusion models with
affine drift.
}
\description{
This package use data augmentation and a Gibbs sampler to sample from the joint posterior
of parameters and augmented data. The main function is 'Estfun' which
produce an object of class 'BIPOD'. See the help page for 'Estfun' for
an example.
}
\details{
\tabular{ll}{
Package: \tab BIPOD\cr
Type: \tab Package\cr
Version: \tab 0.2.1\cr
Date: \tab 2014-03-18\cr
License: \tab GPL-3\cr
}
BBSim
DiffSim
Estfun
plot.BIPOD
ShowModels
}
\author{
Anders Chr. Jensen

Maintainer: Anders Chr. Jensen <anders@math.ku.dk>
}
\references{
  'Importance sampling techniques for estimation of diffusion models' by
  Papaspiliopoulos and Roberts, 'Statistical Methods for Stochastic
  Differential Equations', Monographs on Statistics and Applied
  Probability, Chapman and Hall, 2012 and
  'Markov chain Monte Carlo approach to parameter estimation in the
  FitzHugh-Nagumo model' by Jensen et al, Phys. Rev. E 2012.

}
\keyword{ package }
%\examples{
%% ~~ simple examples of the most important functions ~~
%}
