% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc_plots.R
\name{biPOD_ppc_dens_overlay}
\alias{biPOD_ppc_dens_overlay}
\title{Overlay Posterior Predictive Density}
\usage{
biPOD_ppc_dens_overlay(x, x_fit, n = 500)
}
\arguments{
\item{x}{A \code{bipod} object that contains a 'counts' field with the observed data.}

\item{x_fit}{A fitted model object that contains posterior draws of the replicated data (yrep).}

\item{n}{Number of posterior predictive draws to use in the plot. (default is 500)}
}
\value{
A density overlay plot using \code{bayesplot::ppc_dens_overlay}.
}
\description{
Creates an overlay of the posterior predictive density over the observed data.
This function uses the posterior predictive draws to compare the observed data distribution with the model-predicted distribution.
}
