% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc_plots.R
\name{biPOD_ppc_ribbon}
\alias{biPOD_ppc_ribbon}
\title{Posterior Predictive Ribbon Plot}
\usage{
biPOD_ppc_ribbon(x, x_fit, n = 500, prob = 0.5, prob_outer = 0.9)
}
\arguments{
\item{x}{A \code{bipod} object that contains a 'counts' field with the observed data.}

\item{x_fit}{A fitted model object that contains posterior draws of the replicated data (yrep).}

\item{n}{Number of posterior predictive draws to use in the plot. (default is 500)}

\item{prob}{Numeric value between 0 and 1 representing the probability mass for the inner interval. (default is 0.5)}

\item{prob_outer}{Numeric value between 0 and 1 representing the probability mass for the outer interval. (default is 0.9)}
}
\value{
A ribbon plot using \code{bayesplot::ppc_ribbon}.
}
\description{
Creates a ribbon plot showing the central credible intervals (CI) and outer intervals of the posterior predictive distribution.
The plot visualizes how well the model fits the data across a range of probability intervals.
}
