% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakpoints_inference.R
\name{breakpoints_inference}
\alias{breakpoints_inference}
\title{Fit growth model to bipod object}
\usage{
breakpoints_inference(
  x,
  dt = NULL,
  growth_type = "exponential",
  variational = FALSE,
  spline_spar = 0.5,
  factor_size = 1,
  prior_K = NULL,
  model_selection = FALSE,
  chains = 4,
  iter = 4000,
  cores = 4
)
}
\arguments{
\item{x}{a bipod object}

\item{dt}{half-size of the interval to consider for each proposed breakpoints}

\item{growth_type}{character string specifying the type of growth assumed,
one of "exponential", "logistic".}

\item{variational}{Boolean specifying whether using variational as opposed to mcmc sampling}

\item{spline_spar}{degree of smoothness of the spline to use}

\item{factor_size}{numeric factor by which to divide counts in the bipod object}

\item{prior_K}{Prior mean for the carrying capacity.}

\item{model_selection}{Boolean, if TRUE the best model between exponential and logistic will be used}

\item{chains}{integer number of chains to run in the Markov Chain Monte Carlo (MCMC) algorithm}

\item{iter}{integer number of iterations to run in the MCMC algorithm}

\item{cores}{integer number of cores to use in parallel processing}
}
\value{
the input bipod object with an added 'breakpoints_fit' slot containing the fitted model for the breakpoints
}
\description{
Fit growth model to bipod object
}
