% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_task1.R
\name{fit}
\alias{fit}
\title{Fit a Growth Model to a Bipod Object}
\usage{
fit(
  x,
  growth_type = "exponential",
  infer_t0 = TRUE,
  variational = FALSE,
  factor_size = 1,
  model_selection_algo = "bayes_factor",
  chains = 4,
  iter = 5000,
  cores = 4
)
}
\arguments{
\item{x}{A \code{bipod} object to which the growth model will be fitted.}

\item{growth_type}{A character string specifying the type of growth model to fit.
Options are \code{'exponential'}, \code{'logistic'}, or \code{'both'}. If \code{'both'} is selected,
model selection will be performed to choose the best fit. (default is 'exponential')}

\item{infer_t0}{A logical value indicating whether to infer the time of population origin (\code{t0}).
If \code{TRUE}, the function will estimate this value during model fitting. (default is TRUE)}

\item{variational}{A logical value indicating whether to use variational inference instead of Markov Chain Monte Carlo (MCMC) sampling.
If \code{TRUE}, variational inference is used; otherwise, MCMC sampling is used. (default is FALSE)}

\item{factor_size}{A numeric value representing the factor by which to divide counts in the bipod object.
Must be a positive number and less than or equal to the minimum count in the bipod object. (default is 1)}

\item{model_selection_algo}{A character string specifying the algorithm to use for model selection when \code{growth_type = "both"}.
Options are \code{'bayes_factor'} or \code{'mixture_model'}. (default is "bayes_factor")}

\item{chains}{An integer specifying the number of chains to run in the MCMC algorithm. Ignored if \code{variational = TRUE}. (default is 4)}

\item{iter}{An integer specifying the number of iterations to run in the MCMC algorithm. Ignored if \code{variational = TRUE}. (default is 5000)}

\item{cores}{An integer specifying the number of cores to use for parallel processing during model fitting. (default is 4)}
}
\value{
The input \code{bipod} object with added slots:
\itemize{
\item \code{'fit'}: Contains the fitted model.
\item \code{'fit_info'}: Contains information about the fitting process, including metadata such as sampling type, factor size, growth type, and model selection details.
}
}
\description{
This function fits a specified growth model to a bipod object.
Depending on the input, it can fit an exponential or logistic growth model,
or perform model selection to choose the best fit.
}
