% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit}
\alias{fit}
\title{Fit growth model to bipod object}
\usage{
fit(
  x,
  growth_type = c("exponential", "logistic"),
  model_type = c("gauss", "exact"),
  prior = c("uniform", "invgamma"),
  variational = FALSE,
  factor_size = 1,
  a = 0,
  b = 1,
  g = 1,
  prior_K = NULL,
  chains = 4,
  iter = 4000,
  warmup = 2000,
  cores = 4
)
}
\arguments{
\item{x}{a bipod object}

\item{growth_type}{character string specifying the type of growth assumed,
one of "exponential", "logistic".}

\item{model_type}{character string specifying the model to fit, one of "gauss", "exact".
\itemize{
\item \code{gauss} Use a Gaussian approximation. Works well when counts data are far from 0.
}}

\item{prior}{character string specifying the prior to use, "uniform" or "invagamma".
\itemize{
\item \code{uniform} Uniform U(a,b), where at every iteration you sample with width w = (b-a) / g.
\item \code{invgamma} Inverse gamma prior invGamma(a,b)
}}

\item{variational}{Boolean specifying whether using variational as opposed to mcmc sampling}

\item{factor_size}{numeric factor by which to divide counts in the bipod object}

\item{a}{Minimum value for uniform prior, shape parameter for inverse gamma prior}

\item{b}{Maximum value for uniform prior, scale parameter for inverse gamma prior}

\item{g}{Extra parameters for uniform prior. Sample is done from a uniform centered on the previous
sample and with width w = (b - a) / g.}

\item{prior_K}{Prior mean for the carrying capacity.}

\item{chains}{integer number of chains to run in the Markov Chain Monte Carlo (MCMC) algorithm}

\item{iter}{integer number of iterations to run in the MCMC algorithm}

\item{warmup}{integer number of warmup iterations to run in the MCMC algorithm}

\item{cores}{integer number of cores to use in parallel processing}
}
\value{
the input bipod object with an added 'fit' slot containing the fitted model and an added 'fit_info' slot containing information about the fit
}
\description{
Fit growth model to bipod object
}
