% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_task1.R
\name{fit}
\alias{fit}
\title{Fit growth model to bipod object}
\usage{
fit(
  x,
  growth_type = "exponential",
  variational = FALSE,
  t0_lower_bound = -10,
  factor_size = 1,
  prior_K = NULL,
  model_selection = FALSE,
  chains = 4,
  iter = 4000,
  warmup = 2000,
  cores = 4
)
}
\arguments{
\item{x}{a bipod object}

\item{growth_type}{character string specifying the type of growth assumed,
one of "exponential", "logistic".}

\item{variational}{Boolean specifying whether using variational as opposed to mcmc sampling}

\item{t0_lower_bound}{lower bound of t0, which is the instant of time in which the population is born}

\item{factor_size}{numeric factor by which to divide counts in the bipod object}

\item{prior_K}{Prior mean for the carrying capacity.}

\item{model_selection}{Boolean, if TRUE the best model between exponential and logistic will be used}

\item{chains}{integer number of chains to run in the Markov Chain Monte Carlo (MCMC) algorithm}

\item{iter}{integer number of iterations to run in the MCMC algorithm}

\item{warmup}{integer number of warmup iterations to run in the MCMC algorithm}

\item{cores}{integer number of cores to use in parallel processing}
}
\value{
the input bipod object with an added 'fit' slot containing the fitted model and an added 'fit_info' slot containing information about the fit
}
\description{
Fit growth model to bipod object
}
