% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_task0.R
\name{fit_breakpoints}
\alias{fit_breakpoints}
\title{Fit growth model to bipod object}
\usage{
fit_breakpoints(
  x,
  norm = F,
  n_trials = 5000,
  min_points = 3,
  available_breakpoints = c(1:5),
  constrain_bp_on_x = F
)
}
\arguments{
\item{x}{a bipod object}

\item{norm}{.}

\item{n_trials}{.}

\item{min_points}{.}

\item{available_breakpoints}{.}

\item{constrain_bp_on_x}{.}
}
\value{
the input bipod object with an added 'breakpoints_fit' slot containing the fitted model for the breakpoints
}
\description{
Fit growth model to bipod object
}
