% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_task0.R
\name{fit_breakpoints}
\alias{fit_breakpoints}
\title{Fit growth model to bipod object}
\usage{
fit_breakpoints(
  x,
  norm = TRUE,
  n_trials = 500,
  avg_points_per_window = 3,
  max_breakpoints = 10
)
}
\arguments{
\item{x}{a bipod object}

\item{norm}{.}

\item{n_trials}{.}

\item{avg_points_per_window}{.}

\item{max_breakpoints}{.}
}
\value{
the input bipod object with an added 'breakpoints_fit' slot containing the fitted model for the breakpoints
}
\description{
Fit growth model to bipod object
}
