% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_exp}
\alias{fit_exp}
\title{Fit exponential growth model to bipod object}
\usage{
fit_exp(
  x,
  model_type = c("gauss", "exact"),
  prior = c("uniform", "invgamma"),
  factor_size = 1,
  a = 0,
  b = 1,
  g = 1,
  chains = 4,
  iter = 4000,
  warmup = 2000,
  cores = 4
)
}
\arguments{
\item{x}{a bipod object}

\item{model_type}{character string specifying the model to fit, one of "gauss", "exact".
\itemize{
\item \code{exact} Slowest model. Particularly good for low counts data.
\item \code{gauss} Use a Gaussian approximation. Works well when counts data are far from 0.
}}

\item{prior}{character string specifying the prior to use, "uniform" or "invagamma".
\itemize{
\item \code{uniform} Uniform U(a,b), where at every iteration you sample with width w = (b-a) / g.
\item \code{invgamma} Inverse gamma prior invGamma(a,b)
}}

\item{factor_size}{numeric factor by which to divide counts in the bipod object}

\item{a}{Minimum value for uniform prior, shape parameter for inverse gamma prior}

\item{b}{Maximum value for uniform prior, scale parameter for inverse gamma prior}

\item{g}{Extra parameters for uniform prior. Sample is done from a uniform centered on the previous
sample and with width w = (b - a) / g.}

\item{chains}{integer number of chains to run in the Markov Chain Monte Carlo (MCMC) algorithm}

\item{iter}{integer number of iterations to run in the MCMC algorithm}

\item{warmup}{integer number of warmup iterations to run in the MCMC algorithm}

\item{cores}{integer number of cores to use in parallel processing}
}
\value{
the input bipod object with an added 'fit' slot containing the fitted model and an added 'fit_info' slot containing information about the fit
}
\description{
Fit exponential growth model to bipod object
}
