% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_exp}
\alias{fit_exp}
\title{Fit exponential growth model to bipod object}
\usage{
fit_exp(
  x,
  model_name,
  factor_size = 1,
  fix_rates = 0,
  prior_par = 2,
  chains = 4,
  iter = 4000,
  warmup = 2000,
  cores = 4
)
}
\arguments{
\item{x}{a bipod object}

\item{model_name}{character string specifying the model to fit, one of "gauss", "exact", or "poisson"
\itemize{
\item \code{exact} Slowest model. Particularly good for low counts data.
\item \code{gauss} Use a gaussian approximation. Works well when counts data are far from 0.
\item \code{poisson} Fastest model. Can be applied both to high and low counts data.
}}

\item{factor_size}{numeric factor by which to divide counts in the bipod object}

\item{fix_rates}{logical indicating whether to fix rates at values specified in the bipod object
\itemize{
\item \code{1} Birth rates are free. Death rate is fixed
\item \code{0} Both birth and death rate are fixed.
\item \code{-1} Death rates are free. Birth rate is fixed
}}

\item{prior_par}{numeric parameter for the prior distribution on rates}

\item{chains}{integer number of chains to run in the Markov Chain Monte Carlo (MCMC) algorithm}

\item{iter}{integer number of iterations to run in the MCMC algorithm}

\item{warmup}{integer number of warmup iterations to run in the MCMC algorithm}

\item{cores}{integer number of cores to use in parallel processing}
}
\value{
the input bipod object with an added 'fit' slot containing the fitted model and an added 'fit_info' slot containing information about the fit
}
\description{
Fit exponential growth model to bipod object
}
