% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_log}
\alias{fit_log}
\title{Fit logistic growth model to bipod object}
\usage{
fit_log(
  x,
  model_name,
  factor_size = 1,
  fix_rates = 0,
  prior_par = 2,
  prior_K = NULL,
  chains = 4,
  iter = 4000,
  warmup = 2000,
  cores = 4
)
}
\arguments{
\item{x}{a bipod object}

\item{model_name}{character string specifying the model to fit, currently only "gauss" is supported
\itemize{
\item \code{gauss} Use a gaussian approximation. Works well when counts data are far from 0.
}}

\item{factor_size}{numeric factor by which to divide counts in the bipod object}

\item{fix_rates}{logical indicating whether to fix rates at values specified in the bipod object
\itemize{
\item \code{1} Growth rate is not fixed at every time step.
\item \code{0} Growth rate is fixed throughout all time steps.
}}

\item{prior_par}{numeric parameter for the prior distribution on rates}

\item{prior_K}{numeric parameter for the prior distribution for the carrying capacity,
if NULL a normal prior centered on the highest count will be used}

\item{chains}{integer number of chains to run in the Markov Chain Monte Carlo (MCMC) algorithm}

\item{iter}{integer number of iterations to run in the MCMC algorithm}

\item{warmup}{integer number of warmup iterations to run in the MCMC algorithm}

\item{cores}{integer number of cores to use in parallel processing}
}
\value{
the input bipod object with an added 'fit' slot containing the fitted model and an added 'fit_info' slot containing information about the fit
}
\description{
Fit logistic growth model to bipod object
}
