% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_pop_fit.R
\name{fit_two_pop_model}
\alias{fit_two_pop_model}
\title{Fit a two population growth}
\usage{
fit_two_pop_model(
  x,
  variational = FALSE,
  factor_size = 1,
  chains = 4,
  iter = 4000,
  cores = 4
)
}
\arguments{
\item{x}{a bipod object}

\item{variational}{Boolean specifying whether using variational as opposed to mcmc sampling}

\item{factor_size}{numeric factor by which to divide counts in the bipod object}

\item{chains}{integer number of chains to run in the Markov Chain Monte Carlo (MCMC) algorithm}

\item{iter}{integer number of iterations to run in the MCMC algorithm}

\item{cores}{integer number of cores to use in parallel processing}
}
\value{
the input bipod object with an added 'two_pop_fit' slot containing the fitted model and an added 'two_pop_fit_info' slot containing information about the fit
}
\description{
Fit a two population growth
}
