% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_posterior.R
\name{get_growth_rate_posteriors}
\alias{get_growth_rate_posteriors}
\title{Produces a list of plots, representing the posteriors, one for each growth rate.}
\usage{
get_growth_rate_posteriors(x, labels = NULL)
}
\arguments{
\item{x}{a bipod object with a 'fit' field}

\item{labels}{Vector of labels for the growth rate of each group. If NULL, the standard is 'rho' plus the group index.}
}
\value{
A ggplot object containing the posterior density plots of the growth rates and the prior density plot
}
\description{
Produces a list of plots, representing the posteriors, one for each growth rate.
}
