% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Creates a biPOD object.}
\usage{
init(counts, sample, break_points = NULL)
}
\arguments{
\item{counts}{A dataframe of counts with the following fields:
\itemize{
\item \code{time} time step.
\item \code{count} population count, integer.
}}

\item{sample}{A string containing the sample name.}

\item{break_points}{An array containing the changing points.}
}
\value{
A biPOD object of class \code{bipod}.
}
\description{
Creates a biPOD object from a table containing
population counts ad different time points and sample name.
}
