% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_task2.R
\name{plot_breakpoints_posterior}
\alias{plot_breakpoints_posterior}
\title{Plot the posterior distributions of the inferred breakpoints.}
\usage{
plot_breakpoints_posterior(x, with_histogram = F, alpha = 0.6, colors = NULL)
}
\arguments{
\item{x}{a bipod object with a 'fit' field}

\item{with_histogram}{.}

\item{alpha}{.}

\item{colors}{colors to use for the different inferred breakpoints}
}
\value{
A ggplot object containing the posterior distributions of the inferred breakpoints.
}
\description{
Plot the posterior distributions of the inferred breakpoints.
}
