% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_carrying_capacity_posterior.R
\name{plot_carrying_capacity_posterior}
\alias{plot_carrying_capacity_posterior}
\title{Plot posteriors of carrying capacity K}
\usage{
plot_carrying_capacity_posterior(x, add_prior = F)
}
\arguments{
\item{x}{a bipod object with a 'fit' field}

\item{add_prior}{Boolean, indicate whether to plot also the prior distribution}
}
\value{
A ggplot object containing the posterior density plots of the growth rates and the prior density plot
}
\description{
Plot posteriors of carrying capacity K
}
