% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnostics.R
\name{plot_elbo}
\alias{plot_elbo}
\title{Plot the logarithm of the ELBO and the delta ELBO mean
obtained during the variational sampling.}
\usage{
plot_elbo(x)
}
\arguments{
\item{x}{A biPOD object of class \code{bipod}. Must contains 'fit'}
}
\value{
A ggplot object with traces of the specified parameters
}
\description{
Plot the logarithm of the ELBO and the delta ELBO mean
obtained during the variational sampling.
}
