% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_task1.R
\name{plot_fit}
\alias{plot_fit}
\title{Plot the fit over the input data.}
\usage{
plot_fit(
  x,
  CI = 0.95,
  legend_labels = NULL,
  legend_title = "group",
  zoom = TRUE,
  full_process = TRUE,
  sec_axis_breaks = NULL,
  sec_axis_labels = NULL,
  t0_posterior_color = "darkorange",
  shadows_colors = NULL
)
}
\arguments{
\item{x}{A bipod object of class \code{bipod}. Must contains 'fit'}

\item{CI}{confidence interval for the growth rate to plot}

\item{legend_labels}{Vector containing a name for each unique group in x$counts$group}

\item{legend_title}{Title for the legend. Default is "group"}

\item{zoom}{Boolean. If the plot should present a zoom focused on the observations}

\item{full_process}{Boolean. If the plot should also present the posterior for t0.}

\item{sec_axis_breaks}{Vector containing values in which secondary x axis' breaks will be}

\item{sec_axis_labels}{Vector containing labels for the secondary x axis}

\item{t0_posterior_color}{color to use for the posterior of t0}

\item{shadows_colors}{colors to use for the different time windows}
}
\value{
A plot of the fit over the input data.
}
\description{
Plot the fit over the input data.
}
