% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fits.R
\name{plot_fit}
\alias{plot_fit}
\title{Plot the fit over the input data.}
\usage{
plot_fit(
  x,
  legend_labels = NULL,
  legend_title = "group",
  zoom_limits = NULL,
  sec_axis_breaks = NULL,
  sec_axis_labels = NULL
)
}
\arguments{
\item{x}{A biPOD object of class \code{bipod}. Must contains 'fit'}

\item{legend_labels}{Vector containing a name for each unique group in x$counts$group}

\item{legend_title}{Title for the legend. Default is "group"}

\item{zoom_limits}{Limits of the x-axis for the zoom of the plot}

\item{sec_axis_breaks}{Vector containing values in which secondary x axis' breaks will be}

\item{sec_axis_labels}{Vector containing labels for the secondary x axis}
}
\value{
A plot of the fit over the input data.
}
\description{
Plot the fit over the input data.
}
