% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fits.R
\name{plot_fit}
\alias{plot_fit}
\title{Plot the fit over the input data.}
\usage{
plot_fit(x, final_time = NULL, add_title = F, add_highlights = F, plot_t0 = F)
}
\arguments{
\item{x}{A biPOD object of class \code{bipod}. Must contains 'fit'}

\item{final_time}{The final time for which the fit need to be plotted.}

\item{add_title}{Boolean, indicating whether the plot should have a title}

\item{add_highlights}{Boolean, indicating whether the groups should be highlighted}

\item{plot_t0}{Boolean, indicating whether to plot t0 histogram and whole evolution plot}
}
\value{
A plot of the fit over the input data.
}
\description{
Plot the fit over the input data.
}
