% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_input.R
\name{plot_input}
\alias{plot_input}
\title{Plot the input data of the population over time}
\usage{
plot_input(x, log_scale = F, add_highlights = F)
}
\arguments{
\item{x}{A biPOD object of class \code{bipod}.}

\item{log_scale}{Boolean, indicating whether the plot should have a title}

\item{add_highlights}{Boolean, indicating whether the groups should be highlighted}
}
\value{
A plot. Represents the evolution of the population over time.
}
\description{
Plot the input data of the population over time
}
