% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_selection.R
\name{plot_mixture_model_omega}
\alias{plot_mixture_model_omega}
\title{Plot the fit over the input data.}
\usage{
plot_mixture_model_omega(x, plot_type = "hist", color = "maroon")
}
\arguments{
\item{x}{A biPOD object of class \code{bipod}. Must contains 'fit' and must have been fitted with model selection.}

\item{plot_type}{One between "hist", "violin" and "boxplot"}

\item{color}{Main color of the figure}
}
\value{
A plot of the Bayes Factor with its significance.
}
\description{
Plot the fit over the input data.
}
