% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_posterior.R
\name{plot_normalized_growth_rate_posteriors}
\alias{plot_normalized_growth_rate_posteriors}
\title{Produces a plot of the posteriors for the growth rates.
The posteriors are renormalized such that their highest peaks is at 1.}
\usage{
plot_normalized_growth_rate_posteriors(
  x,
  add_prior = F,
  legend_labels = NULL,
  legend_title = "group"
)
}
\arguments{
\item{x}{a bipod object with a 'fit' field}

\item{add_prior}{Boolean, indicate whether to plot also the prior distribution}

\item{legend_labels}{Vector containing a name for each unique fitted parameters. Default is 'rho'}

\item{legend_title}{Title for the legend. Default is "group"}
}
\value{
A ggplot object containing the posterior density plots of the growth rates and the prior density plot
}
\description{
Produces a plot of the posteriors for the growth rates.
The posteriors are renormalized such that their highest peaks is at 1.
}
