% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_posteriors.R
\name{plot_posterior}
\alias{plot_posterior}
\title{Plot the posterior density for a desired parameter.}
\usage{
plot_posterior(x, x_fit, par_name, color = "black")
}
\arguments{
\item{x}{A biPOD object of class \code{bipod}. Must contains 'fit' and must have been fitted with model selection.}

\item{x_fit}{The fit which contain the desired parameter. It should be inside x.}

\item{par_name}{Name of the desired parameter.}

\item{color}{Color for the density plot.}
}
\value{
A plot of the Bayes Factor with its significance.
}
\description{
Plot the posterior density for a desired parameter.
}
