% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_posteriors.R
\name{plot_posterior}
\alias{plot_posterior}
\title{Plot Posterior Density for a Specific Parameter}
\usage{
plot_posterior(x, x_fit, par_name, color = "black")
}
\arguments{
\item{x}{A \code{bipod} object of class \code{bipod}. Must contain a 'fit' field and should have been fitted with model selection.}

\item{x_fit}{The fit object within \code{x} that contains the desired parameter.}

\item{par_name}{The name of the parameter whose posterior distribution is to be plotted.}

\item{color}{A character string specifying the color to use for the density plot. (default is 'black')}
}
\value{
A \code{ggplot2} object showing the density plot of the specified parameter.
}
\description{
Creates a density plot of the posterior distribution for a specified parameter from a fitted model within a \code{bipod} object.
}
