% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_posteriors.R
\name{plot_posteriors}
\alias{plot_posteriors}
\title{Plot posterior densities for a set of desired parameter.}
\usage{
plot_posteriors(
  x,
  x_fit,
  par_list,
  with_histogram = F,
  alpha = 0.6,
  colors = NULL
)
}
\arguments{
\item{x}{A biPOD object of class \code{bipod}. Must contains 'fit' and must have been fitted with model selection.}

\item{x_fit}{The fit which contain the desired parameter. It should be inside x.}

\item{par_list}{Vector containing names of desired parameters}

\item{with_histogram}{Boolean. If density should overlay histogram of samples.}

\item{alpha}{Alpha value for the plotted densities.}

\item{colors}{Vector containing colors for each parameter.}
}
\value{
A plot of the Bayes Factor with its significance.
}
\description{
Plot posterior densities for a set of desired parameter.
}
