% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_report.R
\name{plot_report}
\alias{plot_report}
\title{Plot traces of specified parameters from a fitted Stan model}
\usage{
plot_report(
  x,
  fit_type = "complex",
  breakpoints_color = "darkgray",
  shadows_colors = NULL,
  t0_posterior_color = "darkorange",
  full_process = FALSE
)
}
\arguments{
\item{x}{A biPOD object of class \code{bipod}. Must contains at least one fit}

\item{fit_type}{.}

\item{breakpoints_color}{.}

\item{shadows_colors}{.}

\item{t0_posterior_color}{.}

\item{full_process}{.}
}
\value{
A ggplot object with traces of the specified parameters
}
\description{
Plot traces of specified parameters from a fitted Stan model
}
