% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fits.R
\name{plot_simple_fit}
\alias{plot_simple_fit}
\title{Plot a simple fit over the input data.}
\usage{
plot_simple_fit(
  x,
  full_process = F,
  legend_labels = NULL,
  legend_title = "group"
)
}
\arguments{
\item{x}{A biPOD object of class \code{bipod}. Must contains 'fit'}

\item{full_process}{Boolean, indicating whether to plot the process starting from t0 or not.}

\item{legend_labels}{Vector containing a name for each unique group in x$counts$group}

\item{legend_title}{Title for the legend. Default is "group"}
}
\value{
A plot of the fit over the input data.
}
\description{
Plot a simple fit over the input data.
}
