% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_task1.R
\name{plot_t0_posterior}
\alias{plot_t0_posterior}
\title{Plot Posteriors of t0}
\usage{
plot_t0_posterior(x, add_prior = F)
}
\arguments{
\item{x}{A \code{bipod} object that contains a 'fit' field.}

\item{add_prior}{Logical value indicating whether to plot the prior distribution along with the posterior distribution. (default is FALSE)}
}
\value{
A \code{ggplot2} object containing the posterior density plot of \code{t0}, and optionally the prior density plot.
}
\description{
Creates a plot of the posterior distribution of the parameter \code{t0} from a \code{bipod} object. Optionally, the prior distribution can be included in the plot.
}
