% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_posterior.R
\name{plot_t0_posterior}
\alias{plot_t0_posterior}
\title{Plot posteriors of t0}
\usage{
plot_t0_posterior(x, add_prior = F)
}
\arguments{
\item{x}{a bipod object with a 'fit' field}

\item{add_prior}{Boolean, indicate whether to plot also the prior distribution}
}
\value{
A ggplot object containing the posterior density plots of the growth rates and the prior density plot
}
\description{
Plot posteriors of t0
}
