% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_general.R
\name{plot_trace_parameters}
\alias{plot_trace_parameters}
\title{Plot MCMC trace plots for parameters}
\usage{
plot_trace_parameters(fit)
}
\arguments{
\item{fit}{A fitted model object containing posterior draws in \code{fit$draws}.}
}
\value{
A trace plot generated by \code{bayesplot::mcmc_trace}.
}
\description{
Generates trace plots for parameters in an MCMC fit.
}
\examples{
\dontrun{
  plot_trace_parameters(fit)
}

}
