% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_two_pop_fit.R
\name{plot_two_pop_fit}
\alias{plot_two_pop_fit}
\title{Plot the fit over the input data.}
\usage{
plot_two_pop_fit(x, CI = 0.95)
}
\arguments{
\item{x}{A bipod object of class \code{bipod}. Must contains 'fit'}

\item{CI}{confidence interval for the growth rate to plot}
}
\value{
A plot of the fit over the input data.
}
\description{
Plot the fit over the input data.
}
