% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_task3.R
\name{plot_two_pop_fit}
\alias{plot_two_pop_fit}
\title{Plot the Fit Over the Input Data for Two Populations}
\usage{
plot_two_pop_fit(
  x,
  CI = 0.95,
  f_posteriors = TRUE,
  t_posteriors = TRUE,
  r_posteriors = TRUE,
  split_process = TRUE,
  resistant_color = "steelblue",
  sensitive_color = "indianred"
)
}
\arguments{
\item{x}{A \code{bipod} object. Must contain a 'two_pop_fit' field.}

\item{CI}{Numeric value representing the confidence interval for the growth rate to plot. (default is 0.95)}

\item{f_posteriors}{.}

\item{t_posteriors}{.}

\item{r_posteriors}{.}

\item{split_process}{Logical value indicating whether to plot the dynamics of the two populations separately. (default is FALSE)}

\item{resistant_color}{.}

\item{sensitive_color}{.}
}
\value{
A \code{ggplot2} object showing the fit over the input data.
}
\description{
Creates a plot of the fit over the input data for a model with two populations.
The plot can display the fit for each population separately and optionally include posteriors for key parameters.
}
