% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ppc_plot}
\alias{ppc_plot}
\title{Plot the posterior predictive checks for the population counts.}
\usage{
ppc_plot(x, prob)
}
\arguments{
\item{x}{A biPOD object of class \code{bipod}. Must contains 'fit'}

\item{prob}{A value between \code{0} and \code{1} indicating the desired probability
mass to include in the intervals of the generated samples.}
}
\value{
A plot. Represents the posterior predictive checks for the population counts.
}
\description{
Plot the posterior predictive checks for the population counts.
}
