% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bipod.R
\name{print.bipod}
\alias{print.bipod}
\title{Print Method for Class \code{'bipod'}}
\usage{
\method{print}{bipod}(x, ...)
}
\arguments{
\item{x}{An object of class \code{bipod}. This object must contain various elements including metadata, counts,
and fitted model objects such as \code{fit}, \code{breakpoints_fit}, and \code{two_pop_fit}.}

\item{...}{Additional parameters for the default S3 method (not used).}
}
\value{
Nothing. This function prints output to the console, providing a detailed summary of the \code{bipod} object.
}
\description{
Custom print method for objects of class \code{bipod}. This method prints a summary of the \code{bipod} object,
including details on the number of observations, the number of time windows, and the status of various tasks
(such as breakpoints inference, two-population inference, and single-population inference).
It also provides summaries of the inferred parameters, including mean values, standard deviations,
quantiles, and Rhat values for each inference task.
}
\details{
The \code{print.bipod} function summarizes key information about the \code{bipod} object, such as:
\itemize{
\item The number of observations and time windows in the dataset.
\item The inference tasks performed, including breakpoints inference, two-population inference, and single-population inference.
\item The Rhat values, mean, standard deviation, and quantiles (5\%, 50\%, 95\%) for the inferred parameters in each task.
\item The status of each task (pass or fail), including the use of variational sampling and any scale factors.
}

The function uses the \code{cli} package to format the output with colors and other visual enhancements.
}
