% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_noisy}
\alias{sim_noisy}
\title{Simulate a deterministic birth-death process for multiple time steps with noise.}
\usage{
sim_noisy(n0, lambda, mu, steps, delta_t, sigma, as_int = T)
}
\arguments{
\item{n0}{Integer value greater than zero. Starting population.}

\item{lambda}{Vector of real values or single real value. Must be greater than zero. Birth rates.}

\item{mu}{Vector of real values or single real value. Must be greater than zero. Death rates.}

\item{steps}{Integer value strictly greater than zero. Number of time steps.}

\item{delta_t}{Vector of real values or single real value. Must be greater than zero. Time intervals.}

\item{sigma}{Real value strictly greater than zero. For each count x, a noise n drawn from a normal
with mean x and variance x * sigma will be added.}

\item{as_int}{Boolean. If noisy counts should be returned as integers.}
}
\value{
A data frame with the columns: "time", "count", "death.rates", and "birth.rates".
}
\description{
Simulate a deterministic birth-death process for multiple time steps with noise.
}
