% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_single_stochastic}
\alias{sim_single_stochastic}
\title{Simulate a single realization of a stochastic process with birth and death events.}
\usage{
sim_single_stochastic(n0, lambda, mu, delta_t)
}
\arguments{
\item{n0}{The initial population size.}

\item{lambda}{The rate at which births occur.}

\item{mu}{The rate at which deaths occur.}

\item{delta_t}{The size of the time step.}
}
\value{
The final population size after the simulation.
}
\description{
This function simulates a single realization of a stochastic process where the
population size is subject to birth and death events. The rates at which these
events occur are given by the \code{lambda} and \code{mu} parameters, respectively. The
function returns the final population size after the simulation is completed.
}
