% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_logistic.R
\name{sim_single_stochastic_logistic}
\alias{sim_single_stochastic_logistic}
\title{Simulate a stochastic birth-death process with logistic grwoth for one time step}
\usage{
sim_single_stochastic_logistic(n0, lambda, mu, K, delta_t)
}
\arguments{
\item{n0}{Integer value greater than zero. Starting population.}

\item{lambda}{Real value greater than zero. Birth rate}

\item{mu}{Real value greater than zero. Death rate}

\item{K}{Integer value, carrying capacity.}

\item{delta_t}{Real value strictly greater than zero. Interval of time.}
}
\value{
An integer number, indicating the population size at the end of the process.
}
\description{
Simulate a stochastic birth-death process with logistic grwoth for one time step
}
