% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim_stochastic}
\alias{sim_stochastic}
\title{Simulate a stochastic birth-death process for multiple time steps}
\usage{
sim_stochastic(n0, lambda, mu, steps, delta_t)
}
\arguments{
\item{n0}{The initial population size.}

\item{lambda}{The rate at which births occur. Can be a single value or a vector
of length \code{steps}.}

\item{mu}{The rate at which deaths occur. Can be a single value or a vector
of length \code{steps}.}

\item{steps}{The number of time steps to simulate.}

\item{delta_t}{The size of the time step. Can be a single value or a vector
of length \code{steps}.}
}
\value{
A tibble with the time, population size, birth rates, and death rates
at each time step.
}
\description{
Simulate a stochastic birth-death process for multiple time steps
}
