% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xenografts}
\alias{xenografts}
\title{Xenografts data}
\format{
\subsection{\code{xenografts}}{

A data frame with 159 rows and 4 columns:
\describe{
\item{mouse}{Mouse id}
\item{tumour_volume}{Tumour volume}
\item{time}{Days, 0 denotes the start of treatment}
\item{treatment_group}{Indicate if the mouse was treated or not}
}
}
}
\source{
.
}
\usage{
xenografts
}
\description{
A subset of data from Sauer et al.
}
\keyword{datasets}
